requirements.txt
torch
torchvision
numpy
pillow
opencv-python
tqdm
ftfy
regex
git+https://github.com/openai/CLIP.git
